backend/requirements.txt
aiohappyeyeballs==2.6.1
redis>=5.0.0
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
bitarray==3.8.0
black==23.12.1
blinker==1.9.0
blis==0.7.11
cachetools==5.5.2
catalogue==2.0.10
certifi==2026.1.4
charset-normalizer==3.4.4
ckzg==1.0.2
click==8.3.1
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.13
cytoolz==1.1.0
datasets==4.5.0
dill==0.3.7
eth-account==0.11.3
eth-hash==0.7.1
eth-keyfile==0.9.1
eth-keys==0.7.0
eth-rlp==1.0.1
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
evaluate==0.4.6
exceptiongroup==1.3.1
fastapi==0.108.0
python-multipart==0.0.22
filelock==3.20.3
flake8==7.0.0
frozenlist==1.8.0
fsspec==2023.10.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hexbytes==0.3.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.7
idna==3.11
importlib-metadata==6.11.0
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kagglehub==1.0.0
kagglesdk==0.1.15
langcodes==3.5.1
lru-dict==1.2.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.15
murmurhash==1.0.15
mypy_extensions==1.1.0
narwhals==2.16.0
networkx==3.4.2
nltk==3.8.1
numpy==2.4.2
packaging==23.2
pandas==3.0.1
parsimonious==0.10.0
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.5.1
pluggy==1.6.0
preshed==3.0.12
propcache==0.4.1
protobuf==4.25.8
py-ecc==8.0.0
pyarrow==23.0.1
pyarrow-hotfix==0.7
pycodestyle==2.11.1
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyflakes==3.2.0
Pygments==2.19.2
PyMuPDF==1.26.7
PyPDF2==3.0.1
pytest==7.4.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.0
python-multipart==0.0.22
pytz==2025.2
pyunormalize==17.0.0
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==13.9.4
rlp==4.1.0
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
seqeval==1.2.2
shellingham==1.5.4
six==1.17.0
smart-open==6.4.0
smmap==5.0.2
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
starlette==0.32.0.post1
streamlit==1.29.0
sympy==1.14.0
tenacity==8.5.0
thinc==8.2.5
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
tomli==2.4.0
toolz==1.1.0
torch==2.10.0  # Install with: pip install torch --index-url https://download.pytorch.org/whl/cpu
tornado==6.5.4
tqdm==4.67.2
transformers==5.0.0
typer==0.9.4
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.25.0
validators==0.35.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.3.4
web3==6.13.0
websockets==16.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
